move taker restart related functionality to swap watcher module
