+++ mm2src/mm2_main/src/lp_swap/swap_watcher.rs
@@ -1,22 +1,28 @@
+use super::taker_swap::TakerSwapCommand;
 use super::{broadcast_p2p_tx_msg, get_payment_locktime, lp_coinfind, taker_payment_spend_deadline, tx_helper_topic,
-            H256Json, SwapsContext, WAIT_CONFIRM_INTERVAL_SEC};
+            AtomicSwap, H256Json, SwapsContext, TakerSavedSwap, TakerSwap, WAIT_CONFIRM_INTERVAL_SEC};
 use crate::mm2::lp_network::{P2PRequestError, P2PRequestResult};
+use crate::mm2::lp_swap::taker_swap::{TakerPaymentSpentData, TakerSavedEvent, TakerSwapEvent};
+use crate::mm2::lp_swap::{SavedSwap, SavedSwapIo, TransactionIdentifier, MAKER_PAYMENT_SPENT_BY_WATCHER_LOG};
 use crate::mm2::MmError;
 use async_trait::async_trait;
 use coins::{CanRefundHtlc, ConfirmPaymentInput, FoundSwapTxSpend, MmCoinEnum, RefundPaymentArgs,
-            SendMakerPaymentSpendPreimageInput, WaitForHTLCTxSpendArgs, WatcherSearchForSwapTxSpendInput,
-            WatcherValidatePaymentInput, WatcherValidateTakerFeeInput};
+            SearchForSwapTxSpendInput, SendMakerPaymentSpendPreimageInput, TransactionEnum, ValidateWatcherSpendInput,
+            WaitForHTLCTxSpendArgs, WatcherSearchForSwapTxSpendInput, WatcherValidatePaymentInput,
+            WatcherValidateTakerFeeInput};
 use common::executor::{AbortSettings, SpawnAbortable, Timer};
 use common::log::{debug, error, info};
 use common::state_machine::prelude::*;
-use common::{now_sec, DEX_FEE_ADDR_RAW_PUBKEY};
+use common::{now_ms, now_sec, DEX_FEE_ADDR_RAW_PUBKEY};
 use futures::compat::Future01CompatExt;
 use mm2_core::mm_ctx::MmArc;
 use mm2_err_handle::prelude::MapToMmResult;
 use mm2_libp2p::{decode_signed, pub_sub_topic, TopicPrefix};
+use rpc::v1::types::Bytes;
 use serde::{Deserialize, Serialize};
 use serde_json as json;
 use std::cmp::min;
+use std::sync::atomic::Ordering;
 use std::sync::Arc;
 use uuid::Uuid;
 
@@ -657,3 +663,285 @@ fn spawn_taker_swap_watcher(ctx: MmArc, watcher_data: TakerSwapWatcherData, veri
 }
 
 pub fn watcher_topic(ticker: &str) -> String { pub_sub_topic(WATCHER_PREFIX, ticker) }
+
+pub async fn get_command_based_on_watcher_activity(
+    swap: &TakerSwap,
+    ctx: &MmArc,
+    mut saved: TakerSavedSwap,
+    command: TakerSwapCommand,
+) -> Result<TakerSwapCommand, String> {
+    #[cfg(not(any(test, feature = "run-docker-tests")))]
+    {
+        let watcher_refund_time = swap.r().data.started_at
+            + (default_watcher_maker_payment_spend_factor() * swap.r().data.lock_duration as f64) as u64;
+        if now_sec() < watcher_refund_time {
+            return Ok(command);
+        }
+    }
+
+    match command {
+        TakerSwapCommand::Start => Ok(command),
+        TakerSwapCommand::Negotiate => Ok(command),
+        TakerSwapCommand::SendTakerFee => Ok(command),
+        TakerSwapCommand::WaitForMakerPayment => Ok(command),
+        TakerSwapCommand::ValidateMakerPayment => Ok(command),
+        TakerSwapCommand::SendTakerPayment => Ok(command),
+        TakerSwapCommand::WaitForTakerPaymentSpend => {
+            match check_taker_payment_spend(swap).await {
+                Ok(Some(FoundSwapTxSpend::Spent(taker_payment_spend_tx))) => {
+                    add_taker_payment_spent_event(swap, &mut saved, &taker_payment_spend_tx).await?;
+                    match check_maker_payment_spend(swap).await {
+                        Ok(Some(FoundSwapTxSpend::Spent(maker_payment_spend_tx))) => {
+                            add_maker_payment_spent_by_watcher_event(saved, ctx, maker_payment_spend_tx).await?;
+                            Ok(TakerSwapCommand::Finish)
+                        },
+                        Ok(Some(FoundSwapTxSpend::Refunded(maker_payment_refund_tx))) => ERR!("Maker has cheated by both spending the taker payment with transaction {:#?}, and refunding the maker payment with transaction {:#?}", taker_payment_spend_tx.tx_hash(), maker_payment_refund_tx.tx_hash()),
+                        Ok(None) => Ok(TakerSwapCommand::SpendMakerPayment),
+                        Err(e) => ERR!("Error {} when trying to find maker payment spend", e),
+                    }
+                },
+                Ok(Some(FoundSwapTxSpend::Refunded(taker_payment_refund_tx))) => {
+                    validate_taker_payment_refund(swap, &taker_payment_refund_tx).await?;
+                    add_taker_payment_refunded_by_watcher_event(saved, ctx, taker_payment_refund_tx).await?;
+                    Ok(TakerSwapCommand::Finish)
+                },
+                Ok(None) => Ok(command),
+                Err(e) => ERR!("Error {} when trying to find taker payment spend", e),
+            }
+        },
+        TakerSwapCommand::SpendMakerPayment => {
+            match check_maker_payment_spend(swap).await {
+                Ok(Some(FoundSwapTxSpend::Spent(maker_payment_spend_tx))) => {
+                    validate_maker_payment_spend(swap, &maker_payment_spend_tx).await?;
+                    add_maker_payment_spent_by_watcher_event(saved, ctx, maker_payment_spend_tx).await?;
+                    Ok(TakerSwapCommand::Finish)
+                },
+                Ok(Some(FoundSwapTxSpend::Refunded(maker_payment_refund_tx))) => ERR!("Maker has cheated by both spending the taker payment, and refunding the maker payment with transaction {:#?}", maker_payment_refund_tx.tx_hash()),
+                Ok(None) => Ok(command),
+                Err(e) => ERR!("Error {} when trying to find maker payment spend", e)
+            }
+        },
+        TakerSwapCommand::PrepareForTakerPaymentRefund |
+        TakerSwapCommand::RefundTakerPayment => {
+            #[cfg(not(any(test, feature = "run-docker-tests")))]
+            {
+                let watcher_refund_time = swap.r().data.started_at + (default_watcher_refund_factor() * swap.r().data.lock_duration as f64) as u64;
+                if now_sec() < watcher_refund_time {
+                    return Ok(command);
+                }
+            }
+
+            let taker_payment_refund_tx = match check_taker_payment_spend(swap).await {
+                Ok(Some(FoundSwapTxSpend::Spent(_))) => return ERR!("Taker payment is not expected to be spent at this point"),
+                Ok(Some(FoundSwapTxSpend::Refunded(taker_payment_refund_tx))) => taker_payment_refund_tx,
+                Ok(None) => return Ok(command),
+                Err(e) => return ERR!("Error {} when trying to find taker payment spend", e),
+            };
+            validate_taker_payment_refund(swap, &taker_payment_refund_tx).await?;
+            add_taker_payment_refunded_by_watcher_event(saved, ctx, taker_payment_refund_tx).await?;
+            Ok(TakerSwapCommand::Finish)
+        },
+        TakerSwapCommand::FinalizeTakerPaymentRefund => Ok(command),
+        TakerSwapCommand::Finish => Ok(command),
+    }
+}
+
+pub async fn check_maker_payment_spend(swap: &TakerSwap) -> Result<Option<FoundSwapTxSpend>, String> {
+    let other_maker_coin_htlc_pub = swap.r().other_maker_coin_htlc_pub;
+    let secret_hash = swap.r().secret_hash.0.clone();
+    let maker_coin_start_block = swap.r().data.maker_coin_start_block;
+    let maker_coin_swap_contract_address = swap.r().data.maker_coin_swap_contract_address.clone();
+
+    let maker_payment = match &swap.r().maker_payment {
+        Some(tx) => tx.tx_hex.0.clone(),
+        None => return ERR!("No info about maker payment, swap is not recoverable"),
+    };
+    let unique_data = swap.unique_swap_data();
+    let watcher_reward = swap.r().watcher_reward;
+
+    swap.maker_coin
+        .search_for_swap_tx_spend_other(SearchForSwapTxSpendInput {
+            time_lock: swap.maker_payment_lock.load(Ordering::Relaxed) as u32,
+            other_pub: other_maker_coin_htlc_pub.as_slice(),
+            secret_hash: &secret_hash,
+            tx: &maker_payment,
+            search_from_block: maker_coin_start_block,
+            swap_contract_address: &maker_coin_swap_contract_address,
+            swap_unique_data: &unique_data,
+            watcher_reward,
+        })
+        .await
+}
+
+pub async fn check_taker_payment_spend(swap: &TakerSwap) -> Result<Option<FoundSwapTxSpend>, String> {
+    let taker_payment = match &swap.r().taker_payment {
+        Some(tx) => tx.tx_hex.0.clone(),
+        None => return ERR!("No info about taker payment, swap is not recoverable"),
+    };
+
+    let other_taker_coin_htlc_pub = swap.r().other_taker_coin_htlc_pub;
+    let taker_coin_start_block = swap.r().data.taker_coin_start_block;
+    let taker_coin_swap_contract_address = swap.r().data.taker_coin_swap_contract_address.clone();
+
+    let taker_payment_lock = match std::env::var("USE_TEST_LOCKTIME") {
+        Ok(_) => swap.r().data.started_at as u32,
+        Err(_) => swap.r().data.taker_payment_lock as u32,
+    };
+    let secret_hash = swap.r().secret_hash.0.clone();
+    let unique_data = swap.unique_swap_data();
+    let watcher_reward = swap.r().watcher_reward;
+
+    swap.taker_coin
+        .search_for_swap_tx_spend_my(SearchForSwapTxSpendInput {
+            time_lock: taker_payment_lock,
+            other_pub: other_taker_coin_htlc_pub.as_slice(),
+            secret_hash: &secret_hash,
+            tx: &taker_payment,
+            search_from_block: taker_coin_start_block,
+            swap_contract_address: &taker_coin_swap_contract_address,
+            swap_unique_data: &unique_data,
+            watcher_reward,
+        })
+        .await
+}
+
+pub async fn validate_maker_payment_spend(
+    swap: &TakerSwap,
+    maker_payment_spend_tx: &TransactionEnum,
+) -> Result<(), String> {
+    let other_maker_coin_htlc_pub = swap.r().other_maker_coin_htlc_pub;
+    let secret_hash = swap.r().secret_hash.0.clone();
+    let maker_coin_swap_contract_address = swap.r().data.maker_coin_swap_contract_address.clone();
+
+    let validate_input = ValidateWatcherSpendInput {
+        payment_tx: maker_payment_spend_tx.tx_hex(),
+        maker_pub: other_maker_coin_htlc_pub.to_vec(),
+        swap_contract_address: maker_coin_swap_contract_address,
+        time_lock: swap.maker_payment_lock.load(Ordering::Relaxed) as u32,
+        secret_hash: secret_hash.clone(),
+        amount: swap.maker_amount.to_decimal(),
+        watcher_reward: None,
+    };
+    swap.maker_coin
+        .taker_validates_maker_payment_spend(validate_input)
+        .compat()
+        .await
+        .map_err(|e| e.to_string())?;
+    Ok(())
+}
+
+pub async fn validate_taker_payment_refund(
+    swap: &TakerSwap,
+    taker_payment_refund_tx: &TransactionEnum,
+) -> Result<(), String> {
+    let other_maker_coin_htlc_pub = swap.r().other_maker_coin_htlc_pub;
+    let taker_coin_swap_contract_address = swap.r().data.taker_coin_swap_contract_address.clone();
+    let taker_payment_lock = match std::env::var("USE_TEST_LOCKTIME") {
+        Ok(_) => swap.r().data.started_at as u32,
+        Err(_) => swap.r().data.taker_payment_lock as u32,
+    };
+    let secret_hash = swap.r().secret_hash.0.clone();
+
+    let validate_input = ValidateWatcherSpendInput {
+        payment_tx: taker_payment_refund_tx.tx_hex(),
+        maker_pub: other_maker_coin_htlc_pub.to_vec(),
+        swap_contract_address: taker_coin_swap_contract_address,
+        time_lock: taker_payment_lock,
+        secret_hash: secret_hash.clone(),
+        amount: swap.taker_amount.to_decimal(),
+        watcher_reward: None,
+    };
+
+    swap.taker_coin
+        .taker_validates_taker_payment_refund(validate_input)
+        .compat()
+        .await
+        .map_err(|e| e.to_string())?;
+    Ok(())
+}
+
+pub async fn add_taker_payment_spent_event(
+    swap: &TakerSwap,
+    saved: &mut TakerSavedSwap,
+    taker_payment_spend_tx: &TransactionEnum,
+) -> Result<(), String> {
+    let secret_hash = swap.r().secret_hash.0.clone();
+    let watcher_reward = swap.r().watcher_reward;
+
+    let tx_hash = taker_payment_spend_tx.tx_hash();
+    info!("Taker payment spend tx {:02x}", tx_hash);
+    let tx_ident = TransactionIdentifier {
+        tx_hex: Bytes::from(taker_payment_spend_tx.tx_hex()),
+        tx_hash,
+    };
+    let secret = match swap
+        .taker_coin
+        .extract_secret(&secret_hash, &tx_ident.tx_hex, watcher_reward)
+        .await
+    {
+        Ok(bytes) => H256Json::from(bytes.as_slice()),
+        Err(_) => {
+            return ERR!("Could not extract secret from taker payment spend transaction");
+        },
+    };
+
+    let event = TakerSwapEvent::TakerPaymentSpent(TakerPaymentSpentData {
+        transaction: tx_ident,
+        secret,
+    });
+    let to_save = TakerSavedEvent {
+        timestamp: now_ms(),
+        event,
+    };
+    saved.events.push(to_save);
+    Ok(())
+}
+
+pub async fn add_maker_payment_spent_by_watcher_event(
+    mut saved: TakerSavedSwap,
+    ctx: &MmArc,
+    maker_payment_spend_tx: TransactionEnum,
+) -> Result<(), String> {
+    let tx_hash = maker_payment_spend_tx.tx_hash();
+    info!("Watcher maker payment spend tx {:02x}", tx_hash);
+    let tx_ident = TransactionIdentifier {
+        tx_hex: Bytes::from(maker_payment_spend_tx.tx_hex()),
+        tx_hash,
+    };
+
+    let event = TakerSwapEvent::MakerPaymentSpentByWatcher(tx_ident);
+    let to_save = TakerSavedEvent {
+        timestamp: now_ms(),
+        event,
+    };
+    saved.events.push(to_save);
+    let new_swap = SavedSwap::Taker(saved);
+    try_s!(new_swap.save_to_db(ctx).await);
+    info!("{}", MAKER_PAYMENT_SPENT_BY_WATCHER_LOG);
+    Ok(())
+}
+
+pub async fn add_taker_payment_refunded_by_watcher_event(
+    mut saved: TakerSavedSwap,
+    ctx: &MmArc,
+    taker_payment_refund_tx: TransactionEnum,
+) -> Result<(), String> {
+    let tx_hash = taker_payment_refund_tx.tx_hash();
+    info!("Taker refund tx hash {:02x}", tx_hash);
+    let tx_ident = TransactionIdentifier {
+        tx_hex: Bytes::from(taker_payment_refund_tx.tx_hex()),
+        tx_hash,
+    };
+
+    let event = TakerSwapEvent::TakerPaymentRefundedByWatcher(Some(tx_ident));
+    let to_save = TakerSavedEvent {
+        timestamp: now_ms(),
+        event,
+    };
+    saved.events.push(to_save);
+
+    let new_swap = SavedSwap::Taker(saved);
+    try_s!(new_swap.save_to_db(ctx).await);
+    info!("Taker payment is refunded by the watcher");
+    Ok(())
+}
